+++ src/experiments/run_mbrl_oracle.py
@@ -4,7 +4,7 @@ import numpy as np
 import torch.multiprocessing as mp
 
 from experiments.experiment_header import get_header
-from environments.cerebral_env_meta.make_env import MakeEnvironment
+from environments.intrepid_env_meta.make_env import MakeEnvironment
 from experiments.experiment_save import terminate
 from learning.core_learner.mbrl_oracle_decoder import MBRLOracleDecoder
