+++ src/experiments/run_sysid.py
@@ -6,7 +6,7 @@ import torch.multiprocessing as mp
 from learning.core_learner.richid import SysID
 from experiments.experiment_save import terminate
 from experiments.experiment_header import get_header
-from environments.cerebral_env_meta.make_env import MakeEnvironment
+from environments.intrepid_env_meta.make_env import MakeEnvironment
 
 
 def main():
