+++ src/learning/core_learner/ppe_util.py
@@ -2,7 +2,7 @@ import os
 import imageio
 import numpy as np
 
-from environments.cerebral_env_meta.environment_keys import EnvKeys
+from environments.intrepid_env_meta.environment_keys import EnvKeys
 from learning.learning_utils.count_probability import CountProbability
 from learning.learning_utils.evaluate_state_decoder import EvaluateStateDecoder
