+++ src/learning/policy_search/greedy_policy_search.py
@@ -6,7 +6,7 @@ import numpy as np
 
 from utils.cuda import cuda_var
 from model.policy.stationary_deterministic_policy import StationaryDeterministicPolicy
-from environments.cerebral_env_meta.environment_wrapper import GenerateEnvironmentWrapper
+from environments.intrepid_env_meta.environment_wrapper import GenerateEnvironmentWrapper
 from learning.policy_search.abstract_policy_search import AbstractPolicySearch
 from learning.learning_utils.contextual_bandit_oracle import ContextualBanditOracle
