+++ src/model/policy/stationary_decoder_dictionary_policy.py
@@ -1,4 +1,4 @@
-from environments.cerebral_env_meta.action_type import ActionType
+from environments.intrepid_env_meta.action_type import ActionType
 from model.policy.stationary_dictionary_policy import StationaryDictionaryPolicy
