+++ src/model/policy/stationary_deterministic_policy.py
@@ -1,7 +1,7 @@
 import torch
 import torch.nn as nn
 
-from environments.cerebral_env_meta.action_type import ActionType
+from environments.intrepid_env_meta.action_type import ActionType
 
 
 class StationaryDeterministicPolicy(nn.Module, ActionType):
