+++ src/model/policy/stationary_dictionary_policy.py
@@ -1,6 +1,6 @@
 import random
 
-from environments.cerebral_env_meta.action_type import ActionType
+from environments.intrepid_env_meta.action_type import ActionType
 
 
 class StationaryDictionaryPolicy(ActionType):
