+++ src/unit_test/make_env.py
@@ -1,4 +1,4 @@
-from environments.cerebral_env_meta.make_env import MakeEnvironment
+from environments.intrepid_env_meta.make_env import MakeEnvironment
 
 
 make_env = MakeEnvironment()
